refactor(#1682 #7): rewrite refs for config_schema/context_builder/python_runner moves
Full sweep (3-class × all-importing-dirs): reyn.config_schema →
reyn.config.config_schema (5), reyn.context_builder →
reyn.core.context_builder (9), reyn.python_runner → reyn.core.python_runner
(4). Zero bare refs remain. No Form-B bare-imports, no __file__ walks,
class-3 clean. ruff clean.
