+++ tests/test_config_mirror_coverage_1056.py
@@ -32,7 +32,7 @@ from __future__ import annotations
 import re
 from pathlib import Path
 
-from reyn.config_schema import walk_config_schema
+from reyn.config.config_schema import walk_config_schema
 
 _REPO_ROOT = Path(__file__).resolve().parents[1]
 _EXAMPLE = _REPO_ROOT / "reyn.local.yaml.example"
