+++ tests/test_config_schema_introspection.py
@@ -19,7 +19,7 @@ import pytest
 import yaml
 
 from reyn.config import ReynConfig
-from reyn.config_schema import (
+from reyn.config.config_schema import (
     MISSING,
     SchemaNode,
     is_valid_config_key,
