+++ tests/test_force_close_phase_activation_1092.py
@@ -21,7 +21,7 @@ from typing import Any
 import pytest
 
 from reyn.chat.router_loop import RouterLoop
-from reyn.context_builder import MAX_OFFLOADED_INLINE_BYTES
+from reyn.core.context_builder import MAX_OFFLOADED_INLINE_BYTES
 from reyn.core.kernel.phase_router_host import PhaseRouterLoopHost
 from reyn.llm.llm import LLMToolCallResult
 from reyn.llm.model_resolver import ModelResolver
