+++ tests/test_offload_read_grant_1383.py
@@ -23,7 +23,7 @@ from pathlib import Path
 
 import pytest
 
-from reyn.context_builder import maybe_ref_artifact, offload_control_ir_result
+from reyn.core.context_builder import maybe_ref_artifact, offload_control_ir_result
 from reyn.security.permissions.permissions import PermissionDecl, PermissionResolver
 from reyn.security.sandbox.policy import SandboxPolicy
